# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Universal AI Tools API
  description: |
    Comprehensive REST API for AI-powered applications using hybrid Go/Rust architecture.
    
    ## Features
    - JWT Authentication
    - Real-time Chat with Local LLMs (Ollama)
    - Vector Search with Qdrant
    - WebSocket Support
    - Agent Orchestration
    - Distributed Tracing
    
    ## Architecture
    - **Go**: API Gateway, WebSocket Service (high concurrency)
    - **Rust**: LLM Router, AI Core (high performance)
    - **TypeScript**: Legacy compatibility layer
    
  version: "1.0.0"
  contact:
    name: Universal AI Tools
    email: support@universal-ai-tools.com
    url: https://docs.universal-ai-tools.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8090
    description: Development server
  - url: https://api.universal-ai-tools.com
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /api/v1/agents:
    get:
      tags: [Agents]
      summary: List all agents
      description: Get a list of all available agents in the system
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, all]
            default: all
        - name: type
          in: query
          schema:
            type: string
            description: Filter by agent type
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 100
      responses:
        '200':
          description: List of agents
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Agent'

    post:
      tags: [Agents]
      summary: Create new agent
      description: Create a new AI agent with specified capabilities
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, type, capabilities]
              properties:
                name:
                  type: string
                  description: Agent name
                type:
                  type: string
                  enum: [conversational, task-based, analytical]
                capabilities:
                  type: array
                  items:
                    type: string
                description:
                  type: string
                config:
                  type: object
                  description: Agent-specific configuration
      responses:
        '201':
          description: Agent created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Agent'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Agent:
      type: object
      properties:
        id:
          type: string
          description: Unique agent identifier
        name:
          type: string
          description: Agent name
        type:
          type: string
          enum: [conversational, task-based, analytical]
        description:
          type: string
          description: Agent description
        capabilities:
          type: array
          items:
            type: string
          description: List of agent capabilities
        status:
          type: string
          enum: [active, inactive, busy, error]
        config:
          type: object
          description: Agent configuration
        createdAt:
          type: string
          format: date-time
        lastActiveAt:
          type: string
          format: date-time
        performance:
          type: object
          properties:
            responseTime:
              type: number
              description: Average response time in milliseconds
            successRate:
              type: number
              description: Success rate as percentage
            totalRequests:
              type: integer
              description: Total number of requests handled

    # Context Schemas
...